LayoutTests/streams/readablestream-cancel-no-chunks.html

<!DOCTYPE html>
<script src="../resources/js-test-pre.js"></script>
<script src="resources/streams-utils.js"></script>
<script>

if (window.testRunner) {
    testRunner.dumpAsText()
    testRunner.waitUntilDone()
}
window.jsTestIsAsync = true;

description('ReadableStream cancel should fulfill promise when cancel before queueing chunk.')

rs = sequentialReadableStream(5);

rs.closed.then(
    function() {
        testPassed('closed promise vended before the cancellation should fulfill');
    },
    function() {
        testFailed('closed promise vended before the cancellation should not be rejected');
        finishJSTest();
    }
  );

rs.ready.then(
    function() {
        testPassed('ready promise vended before the cancellation should fulfill');
    },
    function() {
        testFailed('ready promise vended before the cancellation should not be rejected');
        finishJSTest();
    }
);

rs.cancel();

shouldBeEqualToString('rs.state', 'closed'); // State should be closed.

rs.closed.then(
    function() {
        testPassed('closed promise vended after the cancellation should fulfill');
    },
    function() {
        testFailed('closed promise vended after the cancellation should not be rejected');
        finishJSTest();
    }
);

rs.ready.then(
    function() {
        testPassed('ready promise vended after the cancellation should fulfill');
        finishJSTest();
    },
    function() {
        testFailed('ready promise vended after the cancellation should not be rejected');
        finishJSTest();
    }
);

</script>
<script src="../resources/js-test-post.js"></script>
